Last Update

+++ AppCore.py
@@ -10,6 +10,7 @@ from PySide2.QtCore import QObject, QAbstractListModel, Qt, Slot, Signal, QModel
 from PySide2.QtGui import *
 from numpy import *
 from datetime import datetime
+from TableModel import TableModel
 import time
 import logging
 import subprocess
@@ -27,22 +28,35 @@ os.environ['OPENCV_FFMPEG_CAPTURE_OPTIONS'] = 'rtsp_transport;udp'
 
 DISK = config["main"]["disk"]
 ROOMS_FILE = config["main"]["cameras"]
+SAVE_PATH = config["main"]["savePath"]
 
 
 class FFplayer(QThread):
 
     def __init__(self, parent=None):
         super(FFplayer, self).__init__(parent)
-        self.prev1 = subprocess.Popen(['ffplay', 'rtsp://172.18.191.38/306/3', '-an', '-x', '1280'], stdin=subprocess.PIPE)
+        self.isStart = True
+        self.isChange = False
+        self.currRtsp = ''
+        #'subprocess.Popen(['ffplay', 'rtsp://172.18.191.38/306/3', '-an', '-x', '1280'], stdin=subprocess.PIPE)'
+        self.prev1 = ''
 
 
     def play(self, rtsp):
-        self.prev1.kill()
-        self.prev1 = subprocess.Popen(['ffplay', rtsp, '-an', '-x', '1280'], stdin=subprocess.PIPE)
+        self.currRtsp = rtsp
+        self.isChange = True
+
+
 
     def run(self):
         while True:
-            pass
+            if self.isChange:
+                prev2 = subprocess.Popen(['ffplay', self.currRtsp, '-an', '-x', '1280'], stdin=subprocess.PIPE)
+                time.sleep(4)
+                if self.prev1 != '':
+                    self.prev1.kill()
+                self.prev1 = prev2
+                self.isChange = False
 
 
 
@@ -169,7 +183,7 @@ class SavingCoder(QThread):
     def run(self):
         process_audio = subprocess.Popen(['ffmpeg', '-thread_queue_size', '1024', '-use_wallclock_as_timestamps',
                                           '1', '-rtsp_transport', 'tcp', '-i', self.audio, '-map_metadata', '0',
-                                          '-c', 'copy', 'aud' + self.name], stdin=subprocess.PIPE)
+                                          '-c', 'copy', SAVE_PATH + 'aud' + self.name], stdin=subprocess.PIPE)
 
         while True:
             if self.isRecord == False:
@@ -197,7 +211,7 @@ class SavingStream(QThread):
     def run(self):
         process_video = subprocess.Popen(['ffmpeg', '-thread_queue_size', '1024', '-use_wallclock_as_timestamps',
                                           '1', '-rtsp_transport', 'tcp', '-i', self.rtsp,'-map_metadata', '0', '-map', '0', '-c:v', 'copy', '-an',
-                                          self.name], stdin=subprocess.PIPE)
+                                          SAVE_PATH + self.name], stdin=subprocess.PIPE)
 
 
         while True:
@@ -210,8 +224,8 @@ class SavingStream(QThread):
             print(i)
             time.sleep(1)
 
-        ewq = subprocess.call('ffmpeg -i ' + self.name + ' -i ' + 'aud' + self.audio + ' -c:v ' + 'copy ' + '-c:a ' + 'copy ' +
-                              'final' + self.name, shell=True)
+        ewq = subprocess.call('ffmpeg -i ' + SAVE_PATH + self.name + ' -i ' + SAVE_PATH + 'aud' + self.audio + ' -c:v ' + 'copy ' + '-c:a ' + 'copy ' +
+                              SAVE_PATH + 'final' + self.name, shell=True)
         # path = os.getcwd()
         # os.chdir("../opencast_uploader")
         #os.chdir("..\\opencast_uploader")
@@ -225,40 +239,6 @@ class SavingStream(QThread):
         # # print(self.args[0],self.args[1].split()[0].replace(':','_'),self.args[2],self.args[3],self.args[4], self.args[5], self.args[6], 'final'+self.name)
 
         logger.info('end of recording')
-
-#
-# class SingleStream(QObject):
-#     '''Отдельнй поток под вывод изображения с камеры'''
-#     running = False
-#     newTextAndColor = Signal(ndarray)
-#     rtsp = ""
-#     errors_count = 0
-#     vcap = cv2.VideoCapture('rtsp://172.18.191.54:554/Streaming/Channels/1', cv2.CAP_FFMPEG)
-#     timeflow = 0
-#     def __init__(self, parent=None):
-#         super(SingleStream, self).__init__(parent)
-#
-#     def chngStream(self, str):
-#         # self.timeflow = time.time()
-#         self.rtsp = str
-#         #self.vcap.release()
-#         self.vcap = cv2.VideoCapture(self.rtsp, cv2.CAP_FFMPEG)
-#
-#     # method which will execute algorithm in another thread
-#     def run(self):
-#         while True:
-#             # send signal with new text and color from aonther thread
-#             # self.newTextAndColor.emit(
-#             #     self.common_string
-#             # )
-#             ret, frame = self.vcap.read()
-#             if ret:
-#                 self.newTextAndColor.emit(frame)
-#             # else:
-#             #     self.errors_count = self.errors_count + 1
-#             #     print(self.errors_count, time.time()-self.timeflow)
-#             #     self.vcap.release()
-#             #     self.vcap = cv2.VideoCapture(self.rtsp)
 
 
 
@@ -323,13 +303,72 @@ class VideoModel(QAbstractListModel):
             del self._entries[self.rowCount() - 1]
             self.endRemoveRows()
 
+class QueueConnect(QObject):
+
+    queue_signal = Signal(str, list)
+    def setQueue(self, queue):
+        self.qwe = queue
+
+    def run(self):
+        while True:
+            try:
+                item = self.qwe.get()
+                cmd = item[0]
+                cams = ' '.join(item[1])
+                self.queue_signal.emit(cmd, item[1])
+            except:
+                print('Смерть обработки сообщения API')
+
+            QThread.msleep(2500)
+
+
+
+class SettingsModel(QAbstractListModel):
+    NameRole = Qt.UserRole + 1
+    GenderRole = Qt.UserRole + 2
+    AgeRole = Qt.UserRole + 3
+
+    def __init__(self, data, parent=None):
+        super().__init__(parent)
+        self._data = data
+
+    def rowCount(self, parent=QModelIndex()):
+        return len(self._data)
+
+    def data(self, index, role=Qt.DisplayRole):
+        if index.isValid() and role == Qt.DisplayRole:
+            row = index.row()
+            return f"{self._data[row]['name']} ({self._data[row]['gender']}, {self._data[row]['age']})"
+
+        if role == SettingsModel.NameRole:
+            return self._data[index.row()]['name']
+        elif role == SettingsModel.GenderRole:
+            return self._data[index.row()]['gender']
+        elif role == SettingsModel.AgeRole:
+            return self._data[index.row()]['age']
+
+        return None
+
+    def roleNames(self):
+        return {
+            SettingsModel.NameRole: b'name',
+            SettingsModel.GenderRole: b'gender',
+            SettingsModel.AgeRole: b'age',
+        }
+
+
 
 class AppCore(QObject):
-    def __init__(self, connect, parent=None):
+    def __init__(self, connect, engine, q, parent=None):
         super(AppCore, self).__init__(parent)
         # self.thread = QThread()
+        self.engine = engine
+        freeSpace = self.getFreeSpace()
+        self.engine.rootContext().setContextProperty("freeSpace", freeSpace)
+
         self.data = {}
         self.info = {}
+        self.ipdict = {}
         self.current_room = ''
         self.connector = connect
         self.select_rtsp = []
@@ -342,37 +381,139 @@ class AppCore(QObject):
         #
         self.fplayer = FFplayer()
         self.fplayer.start()
-        # self.player = Player()
-        # self.player.start()
+
+        self.queue = q
+
+        self.thread = QThread()
+        self.queue_an = QueueConnect()
+        self.queue_an.setQueue(self.queue)
+        self.queue_an.moveToThread(self.thread)
+        self.queue_an.queue_signal.connect(self.api_message)
+        self.thread.started.connect(self.queue_an.run)
+        self.thread.start()
+
+        # self.fortqwee = [
+        #     {'name': 'Alice', 'gender': 'female', 'age': 25},
+        #     {'name': 'Bob', 'gender': 'male', 'age': 30},
+        #     {'name': 'Charlie', 'gender': 'male', 'age': 35},
+        # ]
+        # model_setting = SettingsModel(self.fortqwee)
+        # self.engine.rootContext().setContextProperty('settingModel', model_setting)
+
         # cv2.namedWindow("main", cv2.WINDOW_NORMAL)
         # cv2.resizeWindow('main', 900, 900)
         # cv2.setWindowProperty("main", 0, 1)
         # cv2.moveWindow('main', 800, 0)
-        # self.thread = QThread()
+
         # self.singleStream = SingleStream()
         # self.singleStream.moveToThread(self.thread)
         # self.singleStream.newTextAndColor.connect(self.addNewTextAndColor)
         # self.thread.started.connect(self.singleStream.run)
         # self.thread.start()
 
+
+    @Slot(str, list)
+    def api_message(self, cmd, msg):
+        if cmd == 'rec':
+            print('Начало записи:')
+            self.api_record_start(msg)
+
+        elif cmd == 'stop':
+            print('Конец записи')
+            self.api_record_stop()
+            print(msg)
+
+
+
+
+
+    def api_record_stop(self):
+        logger.info('Ending recordings From API')
+        for threads in self.record_threads:
+            threads.stopRecording()
+
+
+    def api_record_start(self, cams):
+        logger.info('Start record from API')
+        string = 'naming'
+        audio_rtsp = ""
+        integer = 1
+        timing = datetime.now()
+        args_aud = [timing.year, timing.day, timing.month, timing.hour, timing.minute]
+        audio_in = {}
+
+        # for id, ad in self.ipdict.items():
+        #     if not ad['room'] in audio_in:
+        #         audio_in[ad['room']] = [ad['audio'][0]['rtsp_main']]
+
+        for cms in cams:
+            if not self.ipdict[cms]['room'] in audio_in:
+                audio_in[self.ipdict[cms]['room']] = [self.ipdict[cms]['audio'][0]['rtsp_main']]
+
+        for key, val in audio_in.items():
+            audio_rtsp = val[0]
+            audio_file = key
+            naming_aud = self.videoNaming(audio_file)
+            audio_in[key].append(naming_aud)
+            audioRecorder = SavingCoder(audio_rtsp, naming_aud, args_aud)
+            audioRecorder.start()
+            self.record_threads.append(audioRecorder)
+
+        for cms in cams:
+            naming = self.videoNaming(self.ipdict[cms]['name'])
+            self.list_files.append(naming)
+            args = [self.ipdict[cms]['room'], self.ipdict[cms]['name'], timing.year, timing.day, timing.month, timing.hour, timing.minute]
+            threadRecord = SavingStream(self.ipdict[cms]['video'], naming, audio_in[self.ipdict[cms]['room']][1], args)
+            threadRecord.start()
+            self.record_threads.append(threadRecord)
+            integer = integer + 1
+
+
+    def makeIPdict(self):
+        '''
+        Создание словаря с камерами и ip-адрессами
+        :return:
+        '''
+        for key, value in self.info.items():
+            aud = ''
+            if value['audio']:
+                aud = value['audio']
+            for cam in value['cameras']:
+                id = (cam['ip'].split('.'))[3]
+                self.ipdict[str(id)] = {'room': key, 'audio': aud, 'video': cam['rtsp_main'], 'name': cam['name']}
+
+        # pprint(self.ipdict)
+
+    @Slot()
+    def printTest(self):
+        print('работает')
+
+
+    @Slot()
+    def getTrigger(self):
+        qwerty = self.engine.rootObjects()[0]
+        qwerty.onTestTrigger()
+
     @Slot(str)
     def getCams(self, str):
         list = []
-        self.select_rtsp.clear()
+        #self.select_rtsp.clear()
 
         self.current_room = str
         for camera in self.info[str]['cameras']:
             list.append({"name": camera['name'],
                          "type": camera['type'],
                          "rtsp": camera['rtsp_main']})
-            self.select_rtsp.append([camera['rtsp_main'],camera['name'], self.current_room])
+            #При переключениии аудиторий все камеры добаляются в запись сразу (ПОка закомментировано)
+            #self.select_rtsp.append([camera['rtsp_main'],camera['name'], self.current_room])
 
 
         if len(self.info[str]['audio']) != 0:
             self.current_audio.clear()
             self.current_audio.append(self.info[str]['audio'][0])
             self.current_audio.append(str)
-        print(self.current_audio)
+
+
         # Изменить отправку на изменение
         self.connector.changeList(list)
 
@@ -380,6 +521,7 @@ class AppCore(QObject):
         with open(ROOMS_FILE, 'r') as file:
             info = json.loads(file.read())
             self.info = info
+            self.makeIPdict()
             return info
 
     def updateCameras(self) -> None:
@@ -412,11 +554,19 @@ class AppCore(QObject):
 
     @Slot(str, str)
     def addSelect(self, rtsp, name) -> None:
-        if [rtsp,name,self.current_room] in self.select_rtsp:
-            self.select_rtsp.remove([rtsp, name, self.current_room])
+
+        if [rtsp,name,self.current_room, self.current_audio[0]['rtsp_main']] in self.select_rtsp:
+            self.select_rtsp.remove([rtsp, name, self.current_room, self.current_audio[0]['rtsp_main']])
+            for cms in self.info[self.current_room]['cameras']:
+                if cms['rtsp_main'] == rtsp:
+                    cms['type'] = 0
         else:
-            self.select_rtsp.append([rtsp, name, self.current_room])
-        print(self.select_rtsp)
+            for cms in self.info[self.current_room]['cameras']:
+                if cms['rtsp_main'] == rtsp:
+                    cms['type'] = 1
+            self.select_rtsp.append([rtsp, name, self.current_room, self.current_audio[0]['rtsp_main']])
+        # print('-'*10)
+        # pprint(self.select_rtsp)
 
     @Slot(str)
     def clearSelected(self, room) -> None:
@@ -433,8 +583,9 @@ class AppCore(QObject):
 
     @Slot(str, str)
     def buttonReact(self, rtsp, name):
-        if (self.isrecord):
-            self.addSelect(rtsp, name)
+        self.addSelect(rtsp, name)
+        print(self.select_rtsp)
+
 
     def getFreeSpace(self) -> str:
         free = psutil.disk_usage(DISK).free / (1024 * 1024 * 1024)
@@ -466,13 +617,17 @@ class AppCore(QObject):
         integer = 1
         timing = datetime.now()
         args_aud = [timing.year, timing.day, timing.month, timing.hour, timing.minute]
+        audio_in = {}
 
-        if self.current_audio:
-            print(self.current_audio)
-            audio_rtsp = self.current_audio[0]['rtsp_main']
-            audio_file = self.current_audio[1]
+        for cms in self.select_rtsp:
+            if not cms[2] in audio_in:
+                audio_in[cms[2]] = [cms[3]]
+
+        for key, val in audio_in.items():
+            audio_rtsp = val[0]
+            audio_file = key
             naming_aud = self.videoNaming(audio_file)
-            self.list_files.append(naming_aud)
+            audio_in[key].append(naming_aud)
             audioRecorder = SavingCoder(audio_rtsp, naming_aud, args_aud)
             audioRecorder.start()
             self.record_threads.append(audioRecorder)
@@ -481,7 +636,7 @@ class AppCore(QObject):
             naming = self.videoNaming(rtsp[1])
             self.list_files.append(naming)
             args = [rtsp[2], rtsp[1], timing.year, timing.day, timing.month, timing.hour, timing.minute]
-            threadRecord = SavingStream(rtsp[0], naming, naming_aud, args)
+            threadRecord = SavingStream(rtsp[0], naming, audio_in[rtsp[2]][1], args)
             threadRecord.start()
             self.record_threads.append(threadRecord)
             integer = integer + 1
